Implement playRound function
This function simulates a round of rock paper scissors. It takes the
human and computer choices as arguments and compares the two choices to
see who has won the round.

Additionally, two new variables have been declared, with each set to
getting the human and computer's choice respectively for the
aforementioned function.