Refactor playRound function
As the roundWinner function was intended to return information about the
round that was played, the playRound function was refactored to remove
code similar to what the roundWinner function was already doing.

Instead, after calculating the player who won the round, the playRound
function will call the roundWinner function and display the returned
information to the user.