Add gameOver function
To display the winner of the game to the user, a gameOver function has
been added.

Additionally, the 'click' event listener for the body element has been
updated so that the gameOver function will be called once a player
reaches a score of 5, in order to determine a winner for the game.